More ES6, no circular dependencies
