+++ rollup.config.js
@@ -1,24 +1,31 @@
-import resolve from '@rollup/plugin-node-resolve'
-import typescript from 'rollup-plugin-typescript2'
-import commonjs from '@rollup/plugin-commonjs'
+import resolve from '@rollup/plugin-node-resolve';
+import typescript from 'rollup-plugin-typescript2';
+import commonjs from '@rollup/plugin-commonjs';
 
-import pkg from './package.json'
+import pkg from './package.json';
 
 export default {
   input: 'src/index.ts',
   output: [
     {
-      banner:'#! /usr/bin/env node\n',
+      banner: '#! /usr/bin/env node\n',
       file: pkg.main,
       format: 'cjs',
     },
     {
-      banner:'#! /usr/bin/env node\n',
+      banner: '#! /usr/bin/env node\n',
       file: pkg.module,
       format: 'es',
     },
   ],
   external: [
+    'child_process',
+    'fs',
+    'https',
+    'os',
+    'path',
+    'tty',
+    'util',
     ...Object.keys(pkg.dependencies || {}),
     ...Object.keys(pkg.peerDependencies || {}),
   ],
@@ -30,4 +37,4 @@ export default {
     }),
     commonjs({ exclude: ['**/__tests__'] }),
   ],
-}
+};
