+++ src/lib/api/tracker/index.ts
@@ -1,3 +1,4 @@
+import { throwError } from '../../utils';
 import { TrackerAppName } from '../interfaces';
 import { Everhour } from './Everhour';
 import { TrackerAPI } from './interfaces';
@@ -8,6 +9,6 @@ export const TrackerFactory = {
       return new Everhour(apiKey);
     }
 
-    throw new Error(`Tracker ${app} is not handled`);
+    throw throwError(`Tracker ${app} is not handled`);
   },
 };
